backend_security/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.4.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.1.1
django-ipware==7.0.1
django-ratelimit==4.1.0
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
python-ipware==3.0.0
redis==5.0.8
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
psycopg2-binary==2.9.9
